@@ -167,7 +167,7 @@
     "chalk": "^2.4.2",
     "clipboard-polyfill": "^2.8.6",
     "compression": "1.7.2",
-    "copy-webpack-plugin": "^5.0.3",
+    "copy-webpack-plugin": "^5.1.1",
     "cross-env": "5.2.0",
     "crypto": "^1.0.1",
     "date-fns": "^1.29.0",
@@ -184,7 +184,7 @@
     "ip": "1.1.5",
     "ipfs-api": "24.0.1",
     "jquery": "3.4.0",
-    "jsbi": "^3.1.1",
+    "jsbi": "^3.1.2",
     "lodash": "4.17.13",
     "markdown-loader": "^5.0.0",
     "minimist": "1.2.0",
@@ -192,9 +192,9 @@
     "object-hash": "^1.3.1",
     "popper.js": "1.14.4",
     "prop-types": "15.6.1",
-    "react": "^16.8.3",
+    "react": "^16.13.1",
     "react-avatar-edit": "^0.8.3",
-    "react-dom": "^16.8.3",
+    "react-dom": "^16.13.1",
     "react-ga": "^2.5.6",
     "react-helmet": "5.2.0",
     "react-intl": "2.4.0",
@@ -202,10 +202,10 @@
     "react-redux": "5.0.7",
     "react-router-dom": "^4.3.1",
     "react-router-redux": "5.0.0-alpha.9",
-    "react-select": "^3.0.4",
+    "react-select": "^3.1.0",
     "react-simplemde-editor": "^3.6.22",
     "redux": "4.0.0",
-    "redux-form": "^7.4.2",
+    "redux-form": "^7.4.3",
     "redux-immutable": "4.0.0",
     "redux-saga": "0.16.0",
     "reselect": "3.0.1",
@@ -281,7 +281,7 @@
     "stylelint-config-recommended": "2.1.0",
     "stylelint-config-styled-components": "0.1.1",
     "stylelint-processor-styled-components": "1.3.1",
-    "svg-inline-loader": "^0.8.0",
+    "svg-inline-loader": "^0.8.2",
     "svg-url-loader": "2.3.2",
     "url-loader": "1.0.1",
     "webpack": "4.12.0",
