@@ -3,8 +3,7 @@
  * AccountProvider
  *
  */
-
-import React from 'react';
+import { useEffect } from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { createStructuredSelector } from 'reselect';
@@ -20,24 +19,24 @@ import { getCurrentAccount } from './actions';
 import { selectLastUpdate } from './selectors';
 import { UPDATE_ACC_PERIOD } from './constants';
 
-/* eslint-disable react/prefer-stateless-function */
-export class AccountProvider extends React.Component {
-  componentDidMount() {
-    this.props.getCurrentAccountDispatch();
-
+export const AccountProvider = ({
+  children,
+  lastUpdate,
+  getCurrentAccountDispatch,
+}) => {
+  useEffect(() => {
+    getCurrentAccountDispatch();
     setInterval(() => {
-      const diff = Date.now() - this.props.lastUpdate;
+      const diff = Date.now() - lastUpdate;
 
       if (diff > UPDATE_ACC_PERIOD) {
-        this.props.getCurrentAccountDispatch();
+        getCurrentAccountDispatch();
       }
     }, UPDATE_ACC_PERIOD / 5);
-  }
+  }, []);
 
-  render() /* istanbul ignore next */ {
-    return this.props.children;
-  }
-}
+  return children;
+};
 
 AccountProvider.propTypes = {
   getCurrentAccountDispatch: PropTypes.func,
@@ -45,25 +44,17 @@ AccountProvider.propTypes = {
   lastUpdate: PropTypes.number,
 };
 
-const mapStateToProps = createStructuredSelector({
-  lastUpdate: selectLastUpdate(),
-});
-
-function mapDispatchToProps(dispatch) /* istanbul ignore next */ {
-  return {
-    getCurrentAccountDispatch: bindActionCreators(getCurrentAccount, dispatch),
-  };
-}
-
 const withConnect = connect(
-  mapStateToProps,
-  mapDispatchToProps,
+  createStructuredSelector({
+    lastUpdate: selectLastUpdate(),
+  }),
+  dispatch => ({
+    getCurrentAccountDispatch: bindActionCreators(getCurrentAccount, dispatch),
+  }),
 );
 
 const withReducer = injectReducer({ key: 'accountProvider', reducer });
 const withSaga = injectSaga({ key: 'accountProvider', saga, mode: DAEMON });
-
-export { mapDispatchToProps };
 
 export default compose(
   withReducer,
